requirements.txt
flask==3.0.0
gunicorn==21.2.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.0.1
mediapipe==0.10.9
onnxruntime==1.19.0
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7
pytest==7.4.4
